requirements.txt
jinja-markdown==1.210911
Jinja2==3.1.6
Markdown==3.10.1
MarkupSafe==3.0.3
Pygments==2.19.2
pymdown-extensions==10.20.1
PyYAML==6.0.3
